requirements.txt
torch
torchvision
mrcfile
starfile
numpy
matplotlib
pandas
lightning
tifffile
h5py
tqdm
hydra-core
omegaconf
jupyter
rootutils
rich
hydra-colorlog
imagecodecs
wandb

